Core/Game Update AnimKitId handling to match TC one
