+++ sql/ashamane/world/2019_08_25_01_world.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `gameobject_template_addon` DROP COLUMN `MovementID`, DROP COLUMN `MeleeID`;
