+++ src/server/game/Entities/Unit/Unit.cpp
@@ -293,7 +293,8 @@ Unit::Unit(bool isWorldObject) :
     i_AI(NULL), i_disabledAI(NULL), m_AutoRepeatFirstCast(false), m_procDeep(0),
     m_removedAurasCount(0), i_motionMaster(new MotionMaster(this)), m_regenTimer(0), m_ThreatManager(this),
     m_vehicle(NULL), m_vehicleKit(NULL), m_unitTypeMask(UNIT_MASK_NONE),
-    m_HostileRefManager(this), _spellHistory(new SpellHistory(this)), _scheduler(this)
+    m_HostileRefManager(this), _aiAnimKitId(0), _movementAnimKitId(0), _meleeAnimKitId(0),
+    _spellHistory(new SpellHistory(this)), _scheduler(this)
 {
     m_objectType |= TYPEMASK_UNIT;
     m_objectTypeId = TYPEID_UNIT;
@@ -11585,21 +11586,26 @@ void Unit::SendDurabilityLoss(Player* receiver, uint32 percent)
     receiver->GetSession()->SendPacket(packet.Write());
 }
 
-void Unit::SetAIAnimKitId(uint16 animKitId, bool oneshot /*= false*/)
+void Unit::PlayOneShotAnimKitId(uint16 animKitId)
 {
-    if (animKitId && !sAnimKitStore.LookupEntry(animKitId))
-        return;
-
-    if (oneshot)
+    if (!sAnimKitStore.LookupEntry(animKitId))
     {
-        WorldPackets::Misc::PlayOneShotAnimKit data;
-        data.Unit = GetGUID();
-        data.AnimKitID = animKitId;
-        SendMessageToSet(data.Write(), true);
+        TC_LOG_ERROR("entities.unit", "Unit::PlayOneShotAnimKitId using invalid AnimKit ID: %u", animKitId);
         return;
     }
 
+    WorldPackets::Misc::PlayOneShotAnimKit data;
+    data.Unit = GetGUID();
+    data.AnimKitID = animKitId;
+    SendMessageToSet(data.Write(), true);
+}
+
+void Unit::SetAIAnimKitId(uint16 animKitId)
+{
     if (_aiAnimKitId == animKitId)
+        return;
+
+    if (animKitId && !sAnimKitStore.LookupEntry(animKitId))
         return;
 
     _aiAnimKitId = animKitId;
