+++ src/server/game/Maps/MapScripts.cpp
@@ -868,7 +868,7 @@ void Map::ScriptsProcess()
             case SCRIPT_COMMAND_PLAY_ANIMKIT:
                 // Source must be Creature.
                 if (Creature* cSource = _GetScriptCreature(source, true, step.script))
-                    cSource->SetAIAnimKitId(step.script->PlayAnimKit.AnimKitID, true);
+                    cSource->PlayOneShotAnimKitId(step.script->PlayAnimKit.AnimKitID);
                 break;
 
             default:
