+++ src/server/scripts/Draenor/Highmaul/boss_tectus.cpp
@@ -317,7 +317,7 @@ class boss_tectus : public CreatureScript
                             return;
 
                         me->SetAIAnimKitId(0);
-                        me->SetAIAnimKitId(eAnimKits::AnimRise, true);
+                        me->PlayOneShotAnimKitId(eAnimKits::AnimRise);
 
                         AddTimedDelayedOperation(4500, [this]() -> void
                         {
@@ -350,7 +350,7 @@ class boss_tectus : public CreatureScript
                         if (m_MoteKilled >= (eMiscs::MotesSpawnCount * 2))
                         {
                             me->SetAIAnimKitId(0);
-                            me->SetAIAnimKitId(eAnimKits::AnimRise, true);
+                            me->PlayOneShotAnimKitId(eAnimKits::AnimRise);
 
                             AddTimedDelayedOperation(4 * TimeConstants::IN_MILLISECONDS, [this]() -> void
                             {
@@ -548,7 +548,7 @@ class boss_tectus : public CreatureScript
                         break;
                     case eSpells::SpawnTectusShards:
                     {
-                        target->SetAIAnimKitId(eAnimKits::AnimRise2, true);
+                        target->PlayOneShotAnimKitId(eAnimKits::AnimRise2);
                         target->RestoreDisplayId();
 
                         ObjectGuid guid = target->GetGUID();
@@ -1103,7 +1103,7 @@ class npc_highmaul_rokka_and_lokk : public CreatureScript
                 {
                     m_Risen = true;
                     me->SetAIAnimKitId(0);
-                    me->SetAIAnimKitId(eAnimKit::AnimRise, true);
+                    me->PlayOneShotAnimKitId(eAnimKit::AnimRise);
 
                     AddTimedDelayedOperation(4500, [this]() -> void
                     {
@@ -1248,7 +1248,7 @@ class npc_highmaul_oro : public CreatureScript
                 {
                     m_Risen = true;
                     me->SetAIAnimKitId(0);
-                    me->SetAIAnimKitId(eAnimKit::AnimRise, true);
+                    me->PlayOneShotAnimKitId(eAnimKit::AnimRise);
 
                     AddTimedDelayedOperation(4500, [this]() -> void
                     {
